Sort out logic to merge feature collection into single feature
From mapbox draw we can get all the currently loaded features
with draw.getAll(). This will return them as a featurecollection
but we ideally want to represent each turf as a single feature,
and if there are multiple shapes merge them as a multipolygon.

So...when we getall from draw we want to merge them into a single
polygon or multipolygon feature.